pyproject.toml
[project]
name = "emailsummarizer"
version = "0.1.0"
description = "Email Summarizer using LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.5.3",
    "exchangelib>=5.5.1",
    "ipykernel>=6.30.1",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "markitdown>=0.1.3",
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "watchdog>=6.0.0",
]
